third_party/blink/web_tests/virtual/threaded-no-composited-antialiasing/animations/inline-block-transform.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<div id="target" style="display:inline-block">x</div>
<script>
var test = async_test('inline-block should run compositor animations');
var anim = target.animate({transform: ['rotate(0deg)', 'rotate(180deg)']}, 100000);
anim.ready.then(() => {
  test.step(() => assert_true(internals.isCompositedAnimation(anim)));
  test.done();
});
</script>
